Game made. Agent framework coded. Card class made.
Created the full game, and it is playable. Created a very simple agent, that can look through its legal cards and pick one using the pick_card function. Also added a glossary for myself, since I'm not used to anglophonic card terms.

Added a couple of methods to CardSet, and abstracted the cards out to its own class, to make suit and value comparisons easier, as well as checking if it's a value card.

This took about 1½ weeks working on it on and off. Maybe overall 6-8 hours. I need to do commits more often.