Urgent bug fix in RandomAgent
I was calling the wrapper function for pick_cards_to_send for the RandomAgent instead of the function that was made to be overwritten, meaning it wouldn't remove the cards it sent to the other players from its hand. I should've commited the change to the send function as a seperate update rather than as a part of the last commit, since it took me ages to track down, because I assumed that it would have something to do with the HumanOpponent features. Definitely a lesson in Git ettiquette.